Refs #3594. Splited integration tests
