+++ test/integration/interaction/Discovery.cpp
@@ -1,0 +1,1 @@
+#include "Discovery.hpp"
